[mtouch] Move the BuildTask classes into their own files.
This is just a pure code move, no functional code changes.
